refactor: 회원가입로직 디버깅 및 코드리팩토링
